added for dept eligibility

+++ src/main/java/com/mahait/gov/in/model/MstEmployeeModel.java
@@ -2,19 +2,20 @@ package com.mahait.gov.in.model;
 
 
 
-import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
 import org.springframework.format.annotation.DateTimeFormat;
 
+import com.mahait.gov.in.entity.OrgPostMst;
+
 import lombok.Data;
 
 @Data
 public class MstEmployeeModel {
 	//Fields
-	private BigInteger employeeId;
+	private Long employeeId;
 	private String sevaarthId;
 	private Character gender;
 	private Long salutation;
